sched/uclamp: Use wrappers for *_{read, write, show} functions
In commit:
    sched/uclamp: Move all tunables to cpusets,

We directly modified and exported *_{read, write, show} functions
to cpuset code.This makes the resulting code quite inconsistent with
the upstream version.

We now use wrappers to reflect the original function, and to preserve
the code in accordance with upstream/mainline.

This is a purely cosmetic change, and no new behaviour should be
exhibited on applying this commit.

(cherry picked from commit bc867bf6c38f72b661bcd0f7006e17c8bfcaa3c3)